+++ lib/Components/CustomWidget/search.dart
@@ -1,0 +1,26 @@
+import 'package:flutter/material.dart';
+
+class SearchBar extends StatelessWidget {
+  const SearchBar({
+    Key key,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      margin: EdgeInsets.symmetric(vertical: 30),
+      padding: EdgeInsets.symmetric(horizontal: 30, vertical: 5),
+      decoration: BoxDecoration(
+        color: Colors.white,
+        borderRadius: BorderRadius.circular(29.5),
+      ),
+      child: TextField(
+        decoration: InputDecoration(
+          hintText: "Search",
+          icon: Icon(Icons.search_outlined),
+          border: InputBorder.none,
+        ),
+      ),
+    );
+  }
+}
